Change NuGet .targets to .props for include order
In order to allow a local build to override the content of NuGet
packages, move the MSBuild properties from a .targets file to a .props
file, which gets included earlier than the project's local overrides.

This allow users to compile locally webrtc.lib and Org.WebRtc.dll and
get the PDBs for debugging, as they are not yet available as NuGet
packages.